+++ src/lib.rs
@@ -5,8 +5,10 @@ pub mod r_cache_config;
 pub mod cache_manager_config;
 pub mod types;
 pub mod errors;
+pub mod async_executor;
 mod cache_task;
 mod metrics;
+mod utils;