+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## 0.3.0 - 2024-12-27
+
+### Added
+
+- Add new invalidation task for replace entry with latest value from remote repository.
+- Add more statistics for Cache's and Cache manager.
+    * Expired tasks
+    * Pending async tasks
+    * Expired async tasks
+    * GET's errors
+- Add Async Executor to be able to execute async cache tasks in Cache Manager.
+- Add integration tests for the new use cases.
+
+### Fixed
+
+
 ## 0.2.0 - 2024-12-24
 
 ### Added